+++ docs/changelog.rst
@@ -24,6 +24,7 @@ Features
 * Add plot options for regression `#1023 <https://github.com/azavea/raster-vision/pull/1023>`_
 * Remove 4GB file size limit from VSI file system, allow streaming reads `#1020 <https://github.com/azavea/raster-vision/pull/1020>`_
 * Add reclassification transformer for segmentation label rasters `#1024 <https://github.com/azavea/raster-vision/pull/1024>`_
+* Allow filtering out chips based on proportion of NODATA pixels `#1025 <https://github.com/azavea/raster-vision/pull/1025>`_
 
 Bug Fixes
 ~~~~~~~~~~~~